requirements.txt
imageio==2.6.1
numpy==1.21.6
Pillow==9.1.0
